change File order and create Select price component
