+++ src/components/Modals/FeedbackModal.js
@@ -45,9 +45,7 @@ const FeedbackModal = props => {
 
   return (
     <div>
-      <p type='button' onClick={handleOpen}>
-        Read full review
-      </p>
+      <p onClick={handleOpen}>Read full review</p>
       <Modal
         aria-labelledby='simple-modal-title'
         aria-describedby='simple-modal-description'
